% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendant.R
\docType{data}
\name{attendant}
\alias{attendant}
\alias{use_attendant}
\alias{attendant_template}
\alias{Attendant}
\title{Attendant}
\format{An object of class \code{shiny.tag} of length 3.}
\usage{
use_attendant(template = attendant_template, start = TRUE)

attendant_template
}
\arguments{
\item{template}{Htmltools tags to shine while content is loading.}

\item{start}{If \code{TRUE} starts on load.}
}
\description{
Create and attendant.
}
\examples{

## ------------------------------------------------
## Method `Attendant$start`
## ------------------------------------------------

\dontrun{Attendant$new()$start()}

## ------------------------------------------------
## Method `Attendant$stop`
## ------------------------------------------------

\dontrun{Attendant$new()$stop()}
}
\keyword{datasets}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-start}{\code{Attendant$start()}}
\item \href{#method-stop}{\code{Attendant$stop()}}
\item \href{#method-clone}{\code{Attendant$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start"></a>}}
\subsection{Method \code{start()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Attendant$start()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Start the attendant.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{Attendant$new()$start()}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\subsection{Method \code{stop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Attendant$stop()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Stop the attendant.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{Attendant$new()$stop()}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Attendant$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
