% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendant.R
\name{attendantBar}
\alias{attendantBar}
\title{Bootstrap 4 Loading Bar}
\usage{
attendantBar(
  id,
  value = 0,
  min = 0,
  max = 100,
  text = NULL,
  color = c("primary", "info", "success", "danger", "warning"),
  striped = FALSE,
  animated = FALSE,
  height = 20,
  class = "",
  style = ""
)
}
\arguments{
\item{id}{A unique identifier for the progress bar.
Used in \code{Attendant} class for handling.}

\item{value, min, max}{Initial value, minimum, and maximum
values the progress bar can take.}

\item{text}{Optional text to display on the progress bar.
This can then be dynamically modified with \code{Attendant}.}

\item{color}{Color of the progress bar.}

\item{striped}{Whether the progress bar should be striped.}

\item{animated}{Whether to animate the stripe on the progress
bar.}

\item{height}{Height of the progress bar, numerical values
are converted to pixels (\code{px} CSS).}

\item{class, style}{Additional style and class to pass to the
parent wrapper of the progress bar.}
}
\description{
Create a Bootstrap 4 progress bar.
}
