% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinners.R
\name{construct}
\alias{construct}
\title{Spinner Constructor}
\usage{
construct(spinner, id)
}
\arguments{
\item{spinner}{Spinner object.}

\item{id}{Id of spinkit to use.}
}
\description{
Construct spinner class.
}
\keyword{internal}
