% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.theme_or_value}
\alias{.theme_or_value}
\title{Theme or Value}
\usage{
.theme_or_value(value, theme_var)
}
\arguments{
\item{value}{Value set by user.}

\item{theme_var}{Theme value to fetch.}
}
\description{
Returns value or theme.
}
\keyword{internal}
