% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hostess.R
\name{hostess}
\alias{hostess}
\alias{use_hostess}
\alias{hostess_loader}
\alias{hostess_init}
\alias{hostess_set}
\alias{Hostess}
\title{Hostess}
\usage{
use_hostess()

hostess_loader(
  id = "hostess",
  preset = NULL,
  text_color = "#FFFFFF",
  with_waiter = TRUE,
  class = "",
  ...
)

hostess_init(id = "hostess")

hostess_set(id = "hostess", value)
}
\arguments{
\item{id}{Id of hostess (valid CSS).}

\item{preset}{A loading bar preset, see section below.}

\item{text_color}{The color of the loading text.}

\item{with_waiter}{Set to \code{FALSE} if used without \code{waiter}.}

\item{class}{CSS class.}

\item{...}{Any other other advanced options to pass to the loaded
see the \href{official documentation}{https://loading.io/progress/}.}

\item{value}{Value to set, between \code{0} and \code{100}.}
}
\description{
Add hostess dependencies.
}
\section{Presets}{

\itemize{
  \item{\code{line}}
  \item{\code{fan}}
  \item{\code{circle}}
  \item{\code{bubble}}
  \item{\code{rainbow}}
  \item{\code{energy}}
  \item{\code{stripe}}
  \item{\code{text}}
}
}

\examples{

## ------------------------------------------------
## Method `Hostess$new`
## ------------------------------------------------

\dontrun{Hostess$new("mySpinner")}

## ------------------------------------------------
## Method `Hostess$set`
## ------------------------------------------------

\dontrun{Hostess$new()}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Hostess$new()}}
\item \href{#method-set}{\code{Hostess$set()}}
\item \href{#method-clone}{\code{Hostess$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hostess$new(id = "hostess")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Id used in `hostess_loader`.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a hostess.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{Hostess$new("mySpinner")}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hostess$set(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Value to set, between \code{0} and \code{100}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set the hostess loading bar.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{Hostess$new()}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hostess$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
