% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr.R
\name{httr_progress}
\alias{httr_progress}
\title{Waitress with httr}
\usage{
httr_progress(object, type = c("down", "up"), pre = NULL, post = NULL)
}
\arguments{
\item{object}{The waitress or attendant object.}

\item{type}{Type of progress to display: either number of
bytes uploaded or downloaded. Passed to \link[httr:progress]{httr::progress}.}

\item{pre, post}{Pre and callback functions to run before
the progress starts or once it is done.}
}
\description{
Use a waitress progress bar with httr requests.
Simply use \code{httr_progress} where you would use
\link[httr:progress]{httr::progress}.
}
\examples{
\dontrun{
cap_speed <- httr::config(max_recv_speed_large = 10000)

httr::GET(
  "http://httpbin.org/bytes/102400", 
  httr_progress(w), 
  cap_speed
)
}

}
