% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lottie.R
\name{lottie}
\alias{lottie}
\title{Lottie}
\usage{
lottie(src, direction = 1L, loop = TRUE, ...)
}
\arguments{
\item{src}{Path to Lottie's JSON file.}

\item{direction}{Direction of animation.}

\item{loop}{Whether to loop the animation.}

\item{...}{Any other options, e.g.: \code{style}.}
}
\description{
Use \href{https://lottiefiles.com}{lottie} animation.
}
