% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholders.R
\name{Placeholder}
\alias{Placeholder}
\title{Placeholders}
\description{
Placeholders

Placeholders
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Placeholder-new}{\code{Placeholder$new()}}
\item \href{#method-Placeholder-show}{\code{Placeholder$show()}}
\item \href{#method-Placeholder-hide}{\code{Placeholder$hide()}}
\item \href{#method-Placeholder-clone}{\code{Placeholder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Placeholder-new"></a>}}
\if{latex}{\out{\hypertarget{method-Placeholder-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Placeholder$new(
  target = ".waiter-placeholder",
  type = c("glow", "wave", "default"),
  session = shiny::getDefaultReactiveDomain()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Placeholder-show"></a>}}
\if{latex}{\out{\hypertarget{method-Placeholder-show}{}}}
\subsection{Method \code{show()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Placeholder$show()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Placeholder-hide"></a>}}
\if{latex}{\out{\hypertarget{method-Placeholder-hide}{}}}
\subsection{Method \code{hide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Placeholder$hide()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Placeholder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Placeholder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Placeholder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
