% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinners.R
\name{preview_spinner}
\alias{preview_spinner}
\title{Preview spinner}
\usage{
preview_spinner(spinner, bg_color = "black")
}
\arguments{
\item{spinner}{A waiter \code{link{spinner}}.}

\item{bg_color}{Background color.}
}
\description{
Allows previewing spinners in web browser or RStudio Viewer.
}
