% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{themes_to_js}
\alias{themes_to_js}
\title{Waitress Themes}
\usage{
themes_to_js(x)
}
\arguments{
\item{x}{Theme to convert.}
}
\description{
Changes waitress themes from R input to js as expected by JS library.
}
\keyword{internal}
