% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waiter.R
\name{transparent}
\alias{transparent}
\title{Transparency}
\usage{
transparent(alpha = 0.4)
}
\arguments{
\item{alpha}{Alpha channel where \code{0} is completely transparent
and \code{1} is opaque.}
}
\description{
A convenience function to create a waiter with transparent background.
}
