% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steward.R
\name{use_steward}
\alias{use_steward}
\title{Steward}
\usage{
use_steward(
  colors = c("#ee7752", "#e73c7e", "#23a6d5", "#23d5ab"),
  speed = 30,
  angle = -45
)
}
\arguments{
\item{colors}{Color palette forming gradient.}

\item{speed}{Seconds it takes to loop over colors.}

\item{angle}{Degrees at which colors slide.}
}
\description{
A colorful steward to work with the \code{\link{waiter}}.
}
