% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendant.R
\name{useAttendant}
\alias{useAttendant}
\title{Attendant Progress Dependencies}
\usage{
useAttendant()
}
\description{
Include in anywhere your shiny UI to import the
dependencies required to run attendant progress.
}
