% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waitress.R
\docType{data}
\name{waitress}
\alias{waitress}
\alias{use_waitress}
\alias{Waitress}
\alias{call_waitress}
\alias{browse_waitresses}
\title{Waitress}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
use_waitress(color = "#b84f3e", percent_color = "#333333")

Waitress

call_waitress(dom = NULL, theme = c("line", "overlay",
  "overlay-radius", "overlay-opacity", "overlay-percent"))

browse_waitresses()
}
\arguments{
\item{color, percent_color}{Color of waitress and color of percent text shown when 
\code{theme} is set to \code{overlay-percent}.}

\item{dom}{Element selector to apply the waitress to, if \code{NULL} then the waitress is applied to the whole screen.}

\item{theme}{A valid theme, see function usage.}

\item{dom}{Element selector to apply the waitress to, if \code{NULL} then the waitress is applied to the whole screen.}

\item{color, percent_color}{Color of waitress and color of percent text shown when 
\code{theme} is set to \code{overlay-percent}.}

\item{theme}{A valid theme, see function usage.}

\item{percent}{Percentage to set waitress to.}

\item{percent}{Percentage to set waitress to.}

\item{ms}{Number of Milliseconds}

\item{percent}{Percentage to increase waitress to.}
}
\description{
Programatically show and hide loading screens.
}
\details{
You can pipe the methods with \code{$}. 
\code{Waitress$new()} and \code{call_waitress()} are equivalent.

Create a waitress.

Start the waitress.

Set the waitress to a specific percentage.

Automatically strat and end the waitress.

Increase the waitress by a percentage.

Hide the waitress.

Print the waitress.
}
\examples{
library(shiny)

ui <- fluidPage(
  use_waitress("red"), # dependencies
  sliderInput("set", "percentage", 1, 100, step = 5, value = 1)
)

server <- function(input, output, session){
 
  w <- Waitress$
   new()$  # call a waitress
  	start() # start waitress
	 
  observeEvent(input$set, {
    w$set(input$set) # set at percentage
  })
}

if(interactive()) shinyApp(ui, server)
 
\dontrun{Waitress$new("#plot")}
\dontrun{Waitress$new("#plot")$start()}
\dontrun{Waitress$new("#plot")$set(20)}
\dontrun{Waitress$new("#plot")$auto(20, 2000)}
\dontrun{Waitress$new("#plot")$increase(30)}
\dontrun{Waitress$new("#plot")$hide()}
\dontrun{Waitress$new("#plot")$hide()}
}
\keyword{datasets}
