% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{withProgressAttendant}
\alias{withProgressAttendant}
\alias{setProgressAttendant}
\alias{incProgressAttendant}
\title{Report Progress Attendant}
\usage{
withProgressAttendant(
  expr,
  ...,
  session = getDefaultReactiveDomain(),
  env = parent.frame(),
  quoted = FALSE
)

setProgressAttendant(
  value = 1,
  text = NULL,
  session = getDefaultReactiveDomain()
)

incProgressAttendant(
  value = 1,
  text = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{expr}{The work to be done.
This expression should contain calls to
setProgressAttendant\code{or}incProgressAttendant`.}

\item{...}{Passed to the Attendant constructor (\code{Attendant$new()}).}

\item{session}{The Shiny session object, as provided by
\code{shinyServer} to the server function. The default is to
automatically find the session by using the current
reactive domain.}

\item{env}{The environment in which \code{expr} should be evaluated.}

\item{quoted}{Whether \code{expr} is a quoted expression
(this is not common).}

\item{value}{Value to set the waitress to or increase it by.}

\item{text}{Text to display on the progress bar.}
}
\description{
Report progress with attendant.
}
