% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{withProgressWaitress}
\alias{withProgressWaitress}
\alias{setProgressWaiter}
\alias{incProgressWaitress}
\title{Report Progress}
\usage{
withProgressWaitress(
  expr,
  ...,
  session = getDefaultReactiveDomain(),
  env = parent.frame(),
  quoted = FALSE
)

setProgressWaiter(value = 1, session = getDefaultReactiveDomain())

incProgressWaitress(value = 1, session = getDefaultReactiveDomain())
}
\arguments{
\item{expr}{The work to be done.
This expression should contain calls to
setProgressWaitress\code{or}incProgressWaitress`.}

\item{...}{Passed to the Waitress constructor (\code{Waitress$new()}).}

\item{session}{The Shiny session object, as provided by
\code{shinyServer} to the server function. The default is to
automatically find the session by using the current
reactive domain.}

\item{env}{The environment in which \code{expr} should be evaluated.}

\item{quoted}{Whether \code{expr} is a quoted expression
(this is not common).}

\item{value}{Value to set the waitress to or increase it by.}
}
\description{
Report progress with waitress.
}
