% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waiter.R
\name{withWaiter}
\alias{withWaiter}
\title{With Waiter}
\usage{
withWaiter(element, html = spin_1(), color = "#333e48", image = "")
}
\arguments{
\item{element}{A reactive element, e.g.: \code{uiOutput},
or \code{plotOutput}.}

\item{html}{HTML content of waiter, generally a spinner, see \code{\link{spinners}}.}

\item{color}{Background color of loading screen.}

\item{image}{Path to background image.}
}
\description{
Adds a waiter to a rective UI element.
Thew waiter is displayed when the element is
invalidated then is removed when the element
receives a new value.
}
