% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hostess.R
\name{hostess}
\alias{hostess}
\alias{use_hostess}
\alias{hostess_loader}
\alias{hostess_init}
\alias{hostess_set}
\alias{Hostess}
\title{Hostess}
\usage{
use_hostess()

hostess_loader(
  id = "hostess",
  preset = NULL,
  text_color = "#FFFFFF",
  center_page = FALSE,
  class = "",
  min = 0,
  max = 100,
  svg = NULL,
  progress_type = c("stroke", "fill"),
  fill_direction = c("ttb", "btt", "ltr", "rtl"),
  stroke_direction = c("normal", "reverse"),
  fill_color = NULL,
  stroke_color = NULL,
  ...,
  with_waiter
)

hostess_init(id = "hostess")

hostess_set(id = "hostess", value)
}
\arguments{
\item{id}{Id of hostess (valid CSS).}

\item{preset}{A loading bar preset, see section below.}

\item{text_color}{The color of the loading text.}

\item{center_page}{By default the hostess is \emph{not} centered in the middle
of the screen, centering in the middle of the page is however ideal when using 
it with waiter full screen, for the latter set to \code{TRUE}.}

\item{class}{CSS class.}

\item{min, max}{Minimum and maximum representing the starting and ending
points of the progress bar.}

\item{svg}{Either an svg path e.g.: \code{M10 10L90 10} or the path to a \code{.svg}
file. Note that if passing the latter it must be made available to Shiny by
placing it either in the \code{www} folder or using \link[shiny]{addResourcePath}.}

\item{progress_type}{The progress type, either \code{stroke} or \code{fill}.
Ther former traces the path of the \code{svg} while the latter fills it progressively.}

\item{fill_direction, stroke_direction}{The direction which the progress bar should
take. Wether \code{fill_direction} or \code{stroke_direction} is used depends on
\code{progress_type}.}

\item{fill_color, stroke_color}{The color to use for the progress bar. 
Wether \code{fill_color} or \code{stroke_color} is used depends on 
\code{progress_type}.}

\item{...}{Any other other advanced options to pass to the loaded
see the \href{https://loading.io/progress/}{official documentation}.}

\item{with_waiter}{Deprecate in favour of \code{center_page}.}

\item{value}{Value to set, between \code{0} and \code{100}.}
}
\description{
Add hostess dependencies.
}
\section{Presets}{

\itemize{
  \item{\code{line}}
  \item{\code{fan}}
  \item{\code{circle}}
  \item{\code{bubble}}
  \item{\code{rainbow}}
  \item{\code{energy}}
  \item{\code{stripe}}
  \item{\code{text}}
}
}

\examples{

## ------------------------------------------------
## Method `Hostess$new`
## ------------------------------------------------

\dontrun{Hostess$new()}

## ------------------------------------------------
## Method `Hostess$set`
## ------------------------------------------------

\dontrun{Hostess$new()$set(20)}

## ------------------------------------------------
## Method `Hostess$inc`
## ------------------------------------------------

\dontrun{Hostess$new()$inc(10)}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{set_loader}}{Set a loader as created by \code{\link{hostess_loader}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Hostess$new()}}
\item \href{#method-start}{\code{Hostess$start()}}
\item \href{#method-print}{\code{Hostess$print()}}
\item \href{#method-set}{\code{Hostess$set()}}
\item \href{#method-inc}{\code{Hostess$inc()}}
\item \href{#method-loader}{\code{Hostess$loader()}}
\item \href{#method-clone}{\code{Hostess$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hostess$new(id = NULL, min = 0, max = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Id used in `hostess_loader` if you generate the loader with 
the \code{loader} method you may leave this \code{NULL}.}

\item{\code{min, max}}{Minimum and maximum representing the starting and ending
points of the progress bar.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a hostess.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{Hostess$new()}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start"></a>}}
\subsection{Method \code{start()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hostess$start()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Start the hostess
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hostess$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Print the hostess
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hostess$set(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Value to set, between \code{0} and \code{100}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set the hostess loading bar.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{Hostess$new()$set(20)}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inc"></a>}}
\subsection{Method \code{inc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hostess$inc(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Value to set, between \code{0} and \code{100}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Increase the hostess loading bar.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{Hostess$new()$inc(10)}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loader"></a>}}
\subsection{Method \code{loader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hostess$loader(
  preset = NULL,
  text_color = "#FFFFFF",
  center_page = FALSE,
  class = "",
  min = NULL,
  max = NULL,
  svg = NULL,
  progress_type = c("stroke", "fill"),
  fill_direction = c("ttb", "btt", "ltr", "rtl"),
  stroke_direction = c("normal", "reverse"),
  fill_color = NULL,
  stroke_color = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{preset}}{A loading bar preset, see section below.}

\item{\code{text_color}}{The color of the loading text.}

\item{\code{center_page}}{By default the hostess is centered in the middle
of the screen, ideal when using it with waiter full screen, set to 
\code{FALSE} to prevent that.}

\item{\code{class}}{CSS class.}

\item{\code{min, max}}{Minimum and maximum representing the starting and ending
points of the progress bar.}

\item{\code{svg}}{Either an svg path e.g.: \code{M10 10L90 10} or the path to a \code{.svg}
file. Note that if passing the latter it must be made available to Shiny by
placing it either in the \code{www} folder or using \link[shiny]{addResourcePath}.}

\item{\code{progress_type}}{The progress type, either \code{stroke} or \code{fill}.
Ther former traces the path of the \code{svg} while the latter fills it progressively.}

\item{\code{fill_direction, stroke_direction}}{The direction which the progress bar should
take. Wether \code{fill_direction} or \code{stroke_direction} is used depends on
\code{progress_type}.}

\item{\code{fill_color, stroke_color}}{The color to use for the progress bar. 
Wether \code{fill_color} or \code{stroke_color} is used depends on 
\code{progress_type}.}

\item{\code{...}}{Any other other advanced options to pass to the loaded
see the \href{https://loading.io/progress/}{official documentation}.}

\item{\code{value}}{Value to set, between \code{0} and \code{100}.}

\item{\code{with_waiter}}{Deprecate in favour of \code{center_page}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a hostess loading bar.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hostess$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
