% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinners.R
\name{spinners}
\alias{spinners}
\alias{spin_rotating_plane}
\alias{spin_fading_circles}
\alias{spin_folding_cube}
\alias{spin_double_bounce}
\alias{spin_wave}
\alias{spin_wandering_cubes}
\alias{spin_pulse}
\alias{spin_chasing_dots}
\alias{spin_three_bounce}
\alias{spin_circle}
\alias{spin_rotate}
\alias{spin_solar}
\alias{spin_orbit}
\alias{spin_squares}
\alias{spin_cube_grid}
\alias{spin_circles}
\alias{spin_orbiter}
\alias{spin_pixel}
\alias{spin_flower}
\alias{spin_dual_ring}
\alias{spin_heart}
\alias{spin_ellipsis}
\alias{spin_facebook}
\alias{spin_hourglass}
\alias{spin_ring}
\alias{spin_ripple}
\alias{browse_spinners}
\alias{browse_waiters}
\title{Spinners}
\usage{
spin_rotating_plane()

spin_fading_circles()

spin_folding_cube()

spin_double_bounce()

spin_wave()

spin_wandering_cubes()

spin_pulse()

spin_chasing_dots()

spin_three_bounce()

spin_circle()

spin_rotate()

spin_solar()

spin_orbit()

spin_squares()

spin_cube_grid()

spin_circles()

spin_orbiter()

spin_pixel()

spin_flower()

spin_dual_ring()

spin_heart()

spin_ellipsis()

spin_facebook()

spin_hourglass()

spin_ring()

spin_ripple()

browse_spinners()

browse_waiters()
}
\description{
Spinkit spinners to use with \code{\link{show_waiter}}.
}
\details{
You can browse the spinners with \code{browse_spinners}.
}
\examples{
if(interactive()) browse_spinners()

}
