% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinners.R
\name{spinners}
\alias{spinners}
\alias{spin_rotating_plane}
\alias{spin_fading_circles}
\alias{spin_folding_cube}
\alias{spin_double_bounce}
\alias{spin_wave}
\alias{spin_wandering_cubes}
\alias{spin_pulse}
\alias{spin_chasing_dots}
\alias{spin_three_bounce}
\alias{spin_circle}
\alias{spin_rotate}
\alias{spin_solar}
\alias{spin_orbit}
\alias{spin_squares}
\alias{spin_cube_grid}
\alias{spin_circles}
\alias{spin_orbiter}
\alias{spin_pixel}
\alias{spin_flower}
\alias{spin_dual_ring}
\alias{spin_heart}
\alias{spin_ellipsis}
\alias{spin_facebook}
\alias{spin_hourglass}
\alias{spin_ring}
\alias{spin_ripple}
\alias{spin_terminal}
\alias{spin_loader}
\alias{spin_throbber}
\alias{spin_refresh}
\alias{spin_heartbeat}
\alias{spin_gauge}
\alias{spin_3k}
\alias{spin_wobblebar}
\alias{spin_atebits}
\alias{spin_whirly}
\alias{spin_flowers}
\alias{spin_dots}
\alias{spin_3circles}
\alias{spin_plus}
\alias{spin_pulsar}
\alias{spin_hexdots}
\alias{spin_inner_circles}
\alias{spin_pong}
\alias{spin_timer}
\alias{spin_ball}
\alias{spin_1}
\alias{spin_2}
\alias{spin_3}
\alias{spin_4}
\alias{spin_5}
\alias{spin_6}
\alias{browse_spinners}
\alias{browse_waiters}
\title{Spinners}
\usage{
spin_rotating_plane()

spin_fading_circles()

spin_folding_cube()

spin_double_bounce()

spin_wave()

spin_wandering_cubes()

spin_pulse()

spin_chasing_dots()

spin_three_bounce()

spin_circle()

spin_rotate()

spin_solar()

spin_orbit()

spin_squares()

spin_cube_grid()

spin_circles()

spin_orbiter()

spin_pixel()

spin_flower()

spin_dual_ring()

spin_heart()

spin_ellipsis()

spin_facebook()

spin_hourglass()

spin_ring()

spin_ripple()

spin_terminal()

spin_loader()

spin_throbber()

spin_refresh()

spin_heartbeat()

spin_gauge()

spin_3k()

spin_wobblebar()

spin_atebits()

spin_whirly()

spin_flowers()

spin_dots()

spin_3circles()

spin_plus()

spin_pulsar()

spin_hexdots()

spin_inner_circles()

spin_pong()

spin_timer()

spin_ball()

spin_1()

spin_2()

spin_3()

spin_4()

spin_5()

spin_6()

browse_spinners()

browse_waiters()
}
\description{
Spinkit spinners to use with \code{\link{show_waiter}}.
}
\details{
You can browse the spinners with \code{browse_spinners}.
}
\examples{
if(interactive()) browse_spinners()

}
