% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinners.R
\name{spinners}
\alias{spinners}
\alias{spin_rotating_plane}
\alias{spin_fading_circles}
\alias{spin_folding_cube}
\alias{spin_double_bounce}
\alias{spin_wave}
\alias{spin_wandering_cubes}
\alias{spin_pulse}
\alias{spin_chasing_dots}
\alias{spin_three_bounce}
\alias{spin_circle}
\alias{spin_cube_grid}
\alias{browse_spinners}
\title{Spinners}
\usage{
spin_rotating_plane()

spin_fading_circles()

spin_folding_cube()

spin_double_bounce()

spin_wave()

spin_wandering_cubes()

spin_pulse()

spin_chasing_dots()

spin_three_bounce()

spin_circle()

spin_cube_grid()

browse_spinners()
}
\description{
Spinkit spinners to use with \code{\link{show_waiter}}.
}
\details{
You can browse the spinners with \code{browse_spinners}.
}
\examples{
if(interactive()) browse_spinners()

}
