% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{waiterOptionInternal}
\alias{waiterOptionInternal}
\alias{.get_waiter_option}
\alias{.get_html}
\alias{.get_color}
\alias{.get_logo}
\alias{.get_image}
\title{Get Waiter Option}
\usage{
.get_waiter_option(theme_var)

.get_html()

.get_color()

.get_logo()

.get_image()
}
\arguments{
\item{theme_var}{Name of option to fetch.}
}
\description{
Returns default waiter value for specific option.
}
\keyword{internal}
