% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waitress.R
\name{waitress}
\alias{waitress}
\alias{use_waitress}
\alias{call_waitress}
\alias{start_waitress}
\alias{set_waitress}
\alias{auto_waitress}
\alias{increase_waitress}
\alias{hide_waitress}
\alias{browse_waitresses}
\title{Waitress}
\usage{
use_waitress(color = "#b84f3e")

call_waitress(dom = NULL, theme = c("line", "overlay",
  "overlay-radius", "overlay-opacity", "overlay-percent"))

start_waitress(waitress)

set_waitress(waitress, percent)

auto_waitress(waitress, percent, ms)

increase_waitress(waitress, percent)

hide_waitress(waitress)

browse_waitresses()
}
\arguments{
\item{color}{Color of waitress.}

\item{dom}{Element selector to apply the waitress to, if \code{NULL} then the waitress is applied to the whole screen.}

\item{theme}{A valid theme, see function usage.}

\item{waitress}{An object of class waitress as returned by \code{call_waitress}.}

\item{percent}{Percentage to set or increase.}

\item{ms}{Milliseconds between each \code{percent} increase.}
}
\description{
Programatically show and hide loading screens.
}
\section{Functions}{

\itemize{
 \item{\code{use_waitress}: Dependencies for waitress, to include anywhere in your UI but ideally at the top.}
 \item{\code{call_waitress}: Initialise.}
 \item{\code{start_waitress}: Start.}
 \item{\code{set_waitress}: Set at specific percentage.}
 \item{\code{auto_waitress}: Auto increment.}
 \item{\code{increase_waitress}: Increase.}
 \item{\code{hide_waitress}: End.}
 \item{\code{browse_waitresses}: Browse waitresses.}
}
}

\examples{
library(shiny)

ui <- fluidPage(
  use_waitress("red"), # dependencies
  sliderInput("set", "percentage", 1, 100, step = 5, value = 1)
)

server <- function(input, output, session){
 
  w1 <- call_waitress() \%>\%  # call a waitress
    start_waitress()

  observeEvent(input$set, {
    set_waitress(w, input$set)
  })
}

if(interactive()) shinyApp(ui, server)
 
}
