% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waitress.R
\name{waitress}
\alias{waitress}
\alias{use_waitress}
\alias{init_waitress}
\alias{start_waitress}
\alias{set_waitress}
\alias{auto_waitress}
\alias{increase_waitress}
\alias{end_waitress}
\title{Waitress}
\usage{
use_waitress()

init_waitress(id = NULL)

start_waitress()

set_waitress(percent)

auto_waitress(percent, ms)

increase_waitress(percent)

end_waitress()
}
\arguments{
\item{id}{Element id, if \code{NULL} then the waitress is applied to the whole screen.}

\item{percent}{Percentage to set or increase.}

\item{ms}{Milliseconds between each \code{percent} increase.}
}
\description{
Programatically show and hide loading screens.
}
\section{Functions}{

\itemize{
 \item{\code{use_waitress}: Dependencies for waitress to include anywhere in your UI but ideally at the top.}
 \item{\code{init_waitress}: Initialise waitress.}
 \item{\code{hide_waiter}: Hide any waiting screen.}
}
}

\examples{
library(shiny)

ui <- fluidPage(
  use_waitress(), # dependencies
  actionButton("init", "Initialise"),
  sliderInput("set", "percentage", 1, 100, step = 5, value = 1)
)

server <- function(input, output, session){
 
  observeEvent(input$init, {
    init_waitress() 
    start_waitress() 
  })

  observeEvent(input$set, {
    set_waitress(input$set)
  })
}

if(interactive()) shinyApp(ui, server)
 
}
