% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waitress.R
\name{waitressClass}
\alias{waitressClass}
\alias{Waitress}
\title{Waitress R6 Class}
\description{
Create a waitress (progress bar) and programmatically set or increase 
its percentage, then hide it when done.
}
\examples{

## ------------------------------------------------
## Method `Waitress$new`
## ------------------------------------------------

\dontrun{Waitress$new("#plot")}

## ------------------------------------------------
## Method `Waitress$start`
## ------------------------------------------------

\dontrun{Waitress$new("#plot")$start()}

## ------------------------------------------------
## Method `Waitress$set`
## ------------------------------------------------

\dontrun{Waitress$new("#plot")$set(20)}

## ------------------------------------------------
## Method `Waitress$auto`
## ------------------------------------------------

\dontrun{Waitress$new("#plot")$auto(20, 2000)}

## ------------------------------------------------
## Method `Waitress$inc`
## ------------------------------------------------

\dontrun{Waitress$new("#plot")$inc(30)}

## ------------------------------------------------
## Method `Waitress$close`
## ------------------------------------------------

\dontrun{Waitress$new("#plot")$close()}

## ------------------------------------------------
## Method `Waitress$print`
## ------------------------------------------------

\dontrun{Waitress$new("#plot")$hide()}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Waitress$new()}}
\item \href{#method-start}{\code{Waitress$start()}}
\item \href{#method-set}{\code{Waitress$set()}}
\item \href{#method-auto}{\code{Waitress$auto()}}
\item \href{#method-increase}{\code{Waitress$increase()}}
\item \href{#method-inc}{\code{Waitress$inc()}}
\item \href{#method-hide}{\code{Waitress$hide()}}
\item \href{#method-close}{\code{Waitress$close()}}
\item \href{#method-print}{\code{Waitress$print()}}
\item \href{#method-clone}{\code{Waitress$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waitress$new(
  selector = NULL,
  theme = c("line", "overlay", "overlay-radius", "overlay-opacity", "overlay-percent"),
  min = 0,
  max = 100,
  infinite = FALSE,
  hide_on_render = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{selector}}{Element selector to apply the waitress to, 
if \code{NULL} then the waitress is applied to the whole screen.}

\item{\code{theme}}{A valid theme, see function usage.}

\item{\code{min, max}}{Minimum and maximum representing the starting and ending
points of the progress bar.}

\item{\code{infinite}}{Set to \code{TRUE} to create a never ending loading bar, ideal
when you cannot compute increments or assess the time it might take before the
loading bar should be removed.}

\item{\code{hide_on_render}}{Set to \code{TRUE} to automatically hide the waitress
when the plot in \code{id} is drawn. Note the latter will only work with
shiny plots, tables, htmlwidgets, etc. but will not work with arbitrary
elements.}

\item{\code{color, percent_color}}{Color of waitress and color of percent text shown when 
\code{theme} is set to \code{overlay-percent}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a waitress.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{Waitress$new("#plot")}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start"></a>}}
\subsection{Method \code{start()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waitress$start(
  html = NULL,
  background_color = "transparent",
  text_color = "black"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{html}}{HTML content to show over the waitress, 
accepts htmltools and shiny tags.}

\item{\code{background_color}}{The background color of the html.}

\item{\code{text_color}}{The color of the \code{html} content.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Start the waitress.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{Waitress$new("#plot")$start()}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waitress$set(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Value to set waitress to.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set the waitress to a specific percentage.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{Waitress$new("#plot")$set(20)}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-auto"></a>}}
\subsection{Method \code{auto()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waitress$auto(value, ms)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Value to set waitress to.}

\item{\code{ms}}{Number of Milliseconds}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Automatically start and end the waitress.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{Waitress$new("#plot")$auto(20, 2000)}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-increase"></a>}}
\subsection{Method \code{increase()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waitress$increase(percent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{percent}}{Percentage to increase waitress to.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Deprecated in favour of \code{inc} method.
Increase the waitress by a percentage.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inc"></a>}}
\subsection{Method \code{inc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waitress$inc(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Value to increase waitress to.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Increase the waitress by a percentage.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{Waitress$new("#plot")$inc(30)}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hide"></a>}}
\subsection{Method \code{hide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waitress$hide()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Deprecated in favour of \code{close} method.
Hide the waitress.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waitress$close()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Close the waitress.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{Waitress$new("#plot")$close()}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waitress$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Print the waitress.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{Waitress$new("#plot")$hide()}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Waitress$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
